requirements.txt
annotated-types==0.7.0
greenlet==3.2.4
mysql-connector-python==9.5.0
pydantic==2.12.4
pydantic_core==2.41.5
PyMySQL==1.1.2
SQLAlchemy==2.0.44
sqlmodel==0.0.27
typing-inspection==0.4.2
typing_extensions==4.15.0
